+++ scripts/reset-repo.sh
@@ -9,7 +9,9 @@ rm -rf example-data \
   data \
   output \
   template.xlsx \
-  env.txt
+  dist
 
 mkdir data
 touch data/.gitkeep
+echo "AP_URL=" > urls.env
+echo "SHACL_URL=" >> urls.env