+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Script to run pipeline locally
+
+### Changed
+
+- GitHub actions uses scripts that are introduced to run pipeline locally
+
 ## [v2.0.0] - 2025-11-25
 
 ### Added